just some renaming and use int for EM reels motorstep like the code expects anyway
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3645 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ dispreel.h
@@ -20,7 +20,7 @@ public:
    int         m_reelcount;         // number of individual reel in the set
    float       m_width, m_height;   // size of each reel
    float       m_reelspacing;       // spacing between each reel and the boarders
-   float       m_motorsteps;        // steps (or frames) to move each reel each frame
+   int         m_motorsteps;        // steps (or frames) to move each reel each frame
    int         m_digitrange;        // max number of digits per reel (usually 9)
 
    char        m_szSound[MAXTOKEN]; // sound to play for each turn of a digit
